requirements/requirements.txt
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-unfold==0.28.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
shortuuid==1.0.13
urllib3==2.2.1
uuid==1.30
uvicorn==0.30.1